requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
pillow==10.2.0
pystray==0.19.5